New folder with the example code moved into it.
Can now choose Merkle hash as well.
