+++ examples/Cargo.toml
@@ -1,40 +1,55 @@
 [package]
-name = "p3-uni-stark"
+name = "p3-examples"
 version = "0.1.0"
 edition = "2021"
 license = "MIT OR Apache-2.0"
 
 [dependencies]
 p3-air.workspace = true
+p3-blake3-air.workspace = true
 p3-field.workspace = true
 p3-challenger.workspace = true
 p3-commit.workspace = true
 p3-dft.workspace = true
+p3-fri.workspace = true
+p3-keccak.workspace = true
+p3-keccak-air.workspace = true
 p3-matrix.workspace = true
 p3-maybe-rayon.workspace = true
+p3-merkle-tree.workspace = true
+p3-monty-31.workspace = true
+p3-poseidon2.workspace = true
+p3-poseidon2-air.workspace = true
+p3-symmetric.workspace = true
+p3-uni-stark.workspace = true
 p3-util.workspace = true
 itertools.workspace = true
-tracing.workspace = true
+rand.workspace = true
 serde = { workspace = true, features = ["derive", "alloc"] }
 
 [dev-dependencies]
 p3-baby-bear.workspace = true
-p3-challenger.workspace = true
-p3-circle.workspace = true
+p3-blake3.workspace = true
 p3-commit = { workspace = true, features = ["test-utils"] }
+p3-circle.workspace = true
+p3-challenger.workspace = true
 p3-dft.workspace = true
-p3-fri.workspace = true
-p3-keccak.workspace = true
+p3-koala-bear.workspace = true
+p3-mds.workspace = true
 p3-matrix.workspace = true
-p3-merkle-tree.workspace = true
 p3-mersenne-31.workspace = true
-p3-symmetric.workspace = true
+p3-sha256.workspace = true
 postcard = { workspace = true, features = ["alloc"] }
-rand.workspace = true
+tracing.workspace = true
+tracing-subscriber = { workspace = true, features = ["std", "env-filter"] }
+tracing-forest = { workspace = true, features = ["ansi", "smallvec"] }
+
 
 [features]
 parallel = ["p3-maybe-rayon/parallel"]
 nightly-features = [
+    "p3-monty-31/nightly-features",
     "p3-baby-bear/nightly-features",
+    "p3-koala-bear/nightly-features",
     "p3-mersenne-31/nightly-features",
 ]
