+++ examples/src/airs.rs
@@ -1,0 +1,67 @@
+use p3_blake3_air::Blake3Air;
+use p3_field::{Field, PrimeField64};
+use p3_keccak_air::KeccakAir;
+use p3_matrix::dense::RowMajorMatrix;
+use p3_poseidon2::GenericPoseidon2LinearLayers;
+use p3_poseidon2_air::VectorizedPoseidon2Air;
+use rand::distributions::Standard;
+use rand::prelude::Distribution;
+
+pub enum ProofGoal<
+    F: Field,
+    LinearLayers,
+    const WIDTH: usize,
+    const SBOX_DEGREE: u64,
+    const SBOX_REGISTERS: usize,
+    const HALF_FULL_ROUNDS: usize,
+    const PARTIAL_ROUNDS: usize,
+    const VECTOR_LEN: usize,
+> {
+    Blake3(Blake3Air),
+    Keccak(KeccakAir),
+    Poseidon2(
+        VectorizedPoseidon2Air<
+            F,
+            LinearLayers,
+            WIDTH,
+            SBOX_DEGREE,
+            SBOX_REGISTERS,
+            HALF_FULL_ROUNDS,
+            PARTIAL_ROUNDS,
+            VECTOR_LEN,
+        >,
+    ),
+}
+
+impl<
+        F: PrimeField64,
+        LinearLayers: GenericPoseidon2LinearLayers<F, WIDTH>,
+        const WIDTH: usize,
+        const SBOX_DEGREE: u64,
+        const SBOX_REGISTERS: usize,
+        const HALF_FULL_ROUNDS: usize,
+        const PARTIAL_ROUNDS: usize,
+        const VECTOR_LEN: usize,
+    >
+    ProofGoal<
+        F,
+        LinearLayers,
+        WIDTH,
+        SBOX_DEGREE,
+        SBOX_REGISTERS,
+        HALF_FULL_ROUNDS,
+        PARTIAL_ROUNDS,
+        VECTOR_LEN,
+    >
+{
+    pub fn generate_trace_rows(&self, num_hashes: usize) -> RowMajorMatrix<F>
+    where
+        Standard: Distribution<F>,
+    {
+        match self {
+            ProofGoal::Blake3(ref b3_air) => b3_air.generate_trace_rows(num_hashes),
+            ProofGoal::Poseidon2(ref p2_air) => p2_air.generate_vectorized_trace_rows(num_hashes),
+            ProofGoal::Keccak(ref k_air) => k_air.generate_trace_rows(num_hashes),
+        }
+    }
+}
